requirements.txt
langchain==0.3.3
langchain-core==0.3.51
langchain-openai==0.2.10
langchain-text-splitters==0.3.0
neo4j==5.27.0
py2neo==2021.2.4
pydantic==2.9.2
RapidFuzz==3.10.0
requests==2.32.3
streamlit==1.40.0
tqdm==4.66.4 
weaviate-client==4.14.1


